afl_custom_trim: Don't allow fuzz inputs to grow
AFLplusplus DOES NOT LIKE IT if you happen to send back
something larger when it asks you to do a trim stage!

This might be a bug with the grammar, but for some grammars,
I have found that the subtree_trimming function can create
a bigger result. This code checks for that case and sends
the original value up instead of allowing a bigger one to
be used.