f1_c_gen.py: Fixup generator randomness
The generator was previously choosing children rules
using some very weird map_rand() calls that appeared to
pretty much always just pick the last rule unless it was
approaching the character limit.

It tended to always pick deeply recursive rules in that case,
resulting in very large strings/numbers/whitespace/etc.

This change causes the rule mapping to work how I *think* it was
intended to work: It picks the rule randomly from the set of
rules that are capable of fitting into the available remaining
characters. This results in behavior like the "Naive generation"
policy from the original Nautilus paper.

The end result is that the random generation rules now favor
much shorter "recursions" in initial seeds. This means that
the mutator will keep the files relatively short (and
therefore fast!) for a while until the recursive mutator extends
these types of values.